# Assembly optimization for AES-NI, and AVX1 and AVX2