… wolfcrypt/benchmark/benchmark.c
… wolfcrypt/src/aes_asm.asm
… wolfcrypt/src/aes_asm.s
… wolfcrypt/test/test.c
… wolfssl/wolfcrypt/cpuid.h
… wolfssl/wolfcrypt/include.am