@toddouska 2017-07-18
AES and SHA-384 improvements look good.  For some reason SHA-512 is about 10% slower for me with these changes:

current
SHA-384    325 MB took 1.007 seconds,  322.868 MB/s Cycles per byte =   7.37
SHA-512    480 MB took 1.007 seconds,  476.506 MB/s Cycles per byte =   4.99

this PR
SHA-384        460 MB took 1.001 seconds,  459.410 MB/s Cycles per byte =   5.18
SHA-512        425 MB took 1.006 seconds,  422.385 MB/s Cycles per byte =   5.63

Can you look into that please?