@SparkiDev 2017-07-18
I believe I've fixed the problem.
Removed extra call to function outside of file that wasn't needed.