@dgarske 2017-07-20
Hi Sean,

GCC 7.1 error:
```
wolfcrypt/src/sha256.c: In function 'Sha256_SetTransform':
wolfcrypt/src/sha256.c:225:9: error: this 'else' clause does not guard... [-Werror=misleading-indentation]
         else
         ^~~~
wolfcrypt/src/sha256.c:227:38: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the 'else'
             Transform_p = Transform; return;
                                      ^~~~~~
```

With speedups:
```
SHA-384        340 MB took 1.006 seconds,  338.029 MB/s Cycles per byte =   7.32
SHA-512        345 MB took 1.004 seconds,  343.670 MB/s Cycles per byte =   7.20
```

Without (stock master):
```
SHA-384    295 MB took 1.010 seconds,  292.006 MB/s Cycles per byte =   8.47
SHA-512    285 MB took 1.003 seconds,  284.130 MB/s Cycles per byte =   8.71
```

```
sysctl -n machdep.cpu.brand_string
Intel(R) Core(TM) i7-3720QM CPU @ 2.60GHz
```

David