@toddouska 2017-07-24
David,

Please do the comparison I did.  --enable-intelmasm on MASTER vs this PR with respect to SHA-512.  You say with speedups and without and your SHA-512 numbers are faster so it looks like this isn't apples to apples.