@dgarske 2017-07-24
@SparkiDev : This was with `--enable-intelasm` only. The results on my Mac with master compared to this PR are not good for SHA-512. Also still getting GCC 7.1 build error (attached possible fix).
<img width="380" alt="screen shot 2017-07-24 at 3 20 47 pm" src="https://user-images.githubusercontent.com/2692870/28547471-d0486fd6-7083-11e7-8462-1d4e69169782.png">


./configure --enable-intelasm

**Intel(R) Core(TM) i7-3720QM CPU @ 2.60GHz (MacBook Pro Retina 2002)**

Master be61d7f:
```
SHA-384    290 MB took 1.016 seconds,  285.522 MB/s Cycles per byte =   8.66
SHA-512    325 MB took 1.014 seconds,  320.622 MB/s Cycles per byte =   7.72
```

This PR:
```
SHA-384        295 MB took 1.013 seconds,  291.128 MB/s Cycles per byte =   8.50
SHA-512        300 MB took 1.000 seconds,  299.988 MB/s Cycles per byte =   8.25
```


**Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz (AVX and AVX2)**
Master be61d7f:
```
SHA-384    365 MB took 1.007 seconds,  362.348 MB/s Cycles per byte =   9.45
SHA-512    495 MB took 1.009 seconds,  490.811 MB/s Cycles per byte =   6.98
```
This PR:
```
SHA-384        490 MB took 1.009 seconds,  485.411 MB/s Cycles per byte =   7.06
SHA-512        490 MB took 1.010 seconds,  485.318 MB/s Cycles per byte =   7.06
```